# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CloudAgent Stage A",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/demo/subjects": {
      "get": {
        "summary": "Subjects",
        "operationId": "subjects_api_v1_demo_subjects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SubjectDTO"
                  },
                  "type": "array",
                  "title": "Response Subjects Api V1 Demo Subjects Get"
                }
              }
            }
          },
          "404": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions": {
      "post": {
        "summary": "Create Session",
        "operationId": "create_session_api_v1_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "subject_id": {
                    "type": "string",
                    "title": "Subject Id",
                    "description": "合成演示身份 ID，可从 /api/v1/demo/subjects 获取"
                  }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "subject_id"
                ],
                "title": "SessionInput"
              },
              "example": {
                "subject_id": "demo-alice"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedSessionDTO"
                }
              }
            }
          },
          "400": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "404": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "409": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "413": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "422": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}": {
      "get": {
        "summary": "Get Session",
        "operationId": "get_session_api_v1_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "x-demo-session-token",
            "in": "header",
            "required": true,
            "description": "会话凭据",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDTO"
                }
              }
            }
          },
          "401": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "403": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "404": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "422": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/close": {
      "post": {
        "summary": "Close Session",
        "operationId": "close_session_api_v1_sessions__session_id__close_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "x-demo-session-token",
            "in": "header",
            "required": true,
            "description": "会话凭据",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedSessionDTO"
                }
              }
            }
          },
          "401": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "403": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "404": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "409": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "422": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{session_id}/runs/stream": {
      "post": {
        "summary": "Stream Run",
        "operationId": "stream_run_api_v1_sessions__session_id__runs_stream_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "x-demo-session-token",
            "in": "header",
            "required": true,
            "description": "会话凭据",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "description": "每次新消息使用的 UUID 幂等键",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "RunEventDTO SSE frames"
                }
              }
            }
          },
          "400": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "401": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "403": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "404": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "409": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "413": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "422": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "429": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "message": {
                    "description": "用户消息；NFC 规范化并裁剪后须为 1 至 4,000 个码点",
                    "title": "Message",
                    "type": "string"
                  }
                },
                "required": [
                  "message"
                ],
                "title": "RunInput",
                "type": "object"
              },
              "example": {
                "message": "产品支持哪些部署版本？"
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-demo-session-token",
            "in": "header",
            "required": true,
            "description": "会话凭据",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDTO"
                }
              }
            }
          },
          "401": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "403": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "404": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "422": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/events": {
      "get": {
        "summary": "Get Events",
        "operationId": "get_events_api_v1_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-demo-session-token",
            "in": "header",
            "required": true,
            "description": "会话凭据",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunEventDTO"
                  },
                  "title": "Response Get Events Api V1 Runs  Run Id  Events Get"
                }
              }
            }
          },
          "401": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "403": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "404": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          },
          "422": {
            "description": "CloudAgent API 错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorDTO": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorDTO"
      },
...
      "ClosedSessionDTO": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "closed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Closed At"
          },
          "close_run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Run Id"
          },
          "terminal_event": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunEventDTO"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "closed_at",
          "close_run_id",
          "terminal_event"
        ],
        "title": "ClosedSessionDTO"
      },
      "CreatedSessionDTO": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "subject_id": {
            "type": "string",
            "title": "Subject Id"
          },
          "subject_label": {
            "type": "string",
            "title": "Subject Label"
          },
          "demo_session_token": {
            "type": "string",
            "title": "Demo Session Token"
          },
          "token_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Token Expires At"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "subject_id",
          "subject_label",
          "demo_session_token",
          "token_expires_at",
          "source"
        ],
        "title": "CreatedSessionDTO"
      },
...
      "MessageDTO": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content",
          "source",
          "run_id"
        ],
        "title": "MessageDTO"
      },
...
      "RunDTO": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "intent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent"
          },
          "agent_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Path"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "session_id",
          "status",
          "intent",
          "agent_path",
          "error_code",
          "started_at",
          "finished_at"
        ],
        "title": "RunDTO"
      },
      "RunEventDTO": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "run_id",
          "sequence",
          "created_at",
          "type",
          "payload"
        ],
        "title": "RunEventDTO"
      },
      "SessionDTO": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "subject_id": {
            "type": "string",
            "title": "Subject Id"
          },
          "token_expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Token Expires At"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageDTO"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "subject_id",
          "token_expires_at",
          "closed_at",
          "messages"
        ],
        "title": "SessionDTO"
      },
      "SubjectDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "role"
        ],
        "title": "SubjectDTO"
      },
...
    }
  }
}